requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
fastapi==0.75.2
frozenlist==1.3.0
future==0.18.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
install==1.3.5
line-bot-sdk==2.2.1
multidict==6.0.2
pydantic==1.9.0
requests==2.27.1
sniffio==1.2.0
starlette==0.17.1
typing==3.7.4.3
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
yarl==1.7.2
